Move jdk.thread under the 'jdk' package
