+++ agent/agent-plugins/jdk-thread/src/main/java/org/bithon/agent/plugin/jdk/thread/interceptor/ThreadPoolExecutor$Ctor.java
@@ -14,16 +14,16 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.thread.interceptor;
+package org.bithon.agent.plugin.jdk.thread.interceptor;
 
 import org.bithon.agent.instrumentation.aop.context.AopContext;
 import org.bithon.agent.instrumentation.aop.interceptor.InterceptionDecision;
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.AroundInterceptor;
 import org.bithon.agent.instrumentation.expt.AgentException;
-import org.bithon.agent.plugin.thread.metrics.ThreadPoolExecutorMetrics;
-import org.bithon.agent.plugin.thread.metrics.ThreadPoolMetricRegistry;
-import org.bithon.agent.plugin.thread.utils.ObservedExecutionHandler;
-import org.bithon.agent.plugin.thread.utils.ThreadPoolNameHelper;
+import org.bithon.agent.plugin.jdk.thread.metrics.ThreadPoolExecutorMetrics;
+import org.bithon.agent.plugin.jdk.thread.metrics.ThreadPoolMetricRegistry;
+import org.bithon.agent.plugin.jdk.thread.utils.ObservedExecutionHandler;
+import org.bithon.agent.plugin.jdk.thread.utils.ThreadPoolNameHelper;
 import org.bithon.component.commons.logging.LoggerFactory;
 
 import java.util.concurrent.BlockingQueue;
